+++ src/main/java/com/ccor/ecommerce/model/dto/ProductSoldRequestDTO.java
@@ -4,6 +4,6 @@ public record ProductSoldRequestDTO(
         String barCode,
         String name,
         int amount,
-        double price
+        int price
 ) {
 }
