+++ src/main/java/com/ccor/ecommerce/model/dto/ProductStockRequestDTO.java
@@ -1,10 +1,13 @@
 package com.ccor.ecommerce.model.dto;
 
+
+import org.springframework.web.multipart.MultipartFile;
+
 public record ProductStockRequestDTO(
         String name,
         int amount,
-        double pricePerUnit,
+        int pricePerUnit,
         String barCode,
         boolean enableProduct
-) {
-}
+) {}
+
